Look up the historical exchange rate between a base and target currency. Defaults to USD - HUF.

Look up the historical exchange rate between a base and target currency. Defaults to USD - HUF.

get_exchange_rates(
  base_currency = "USD",
  target_currency = "HUF",
  start_date = Sys.Date() - 30,
  end_date = Sys.Date(),
  retried = 0
)

Arguments

base_currency
currency symbol

target_currency
currency symbol

start_date
date

end_date
date

retried
number of times the function has been retried

Value

data.table object

Examples

get_exchange_rates()

#>           date     rate
#>  1: 2021-04-27 299.9649
#>  2: 2021-04-28 298.3900
#>  3: 2021-04-29 297.3022
#>  4: 2021-04-30 299.5300
#>  5: 2021-05-01 299.6400
#>  6: 2021-05-02 299.3500
#>  7: 2021-05-03 298.1286
#>  8: 2021-05-04 299.7200
#>  9: 2021-05-05 298.9222
#> 10: 2021-05-06 297.1086
#> 11: 2021-05-07 294.7075
#> 12: 2021-05-08 294.6800
#> 13: 2021-05-09 294.1630
#> 14: 2021-05-10 294.6985
#> 15: 2021-05-11 295.3076
#> 16: 2021-05-12 296.0395
#> 17: 2021-05-13 295.4357
#> 18: 2021-05-14 292.4401
#> 19: 2021-05-15 292.4400
#> 20: 2021-05-16 292.6000
#> 21: 2021-05-17 289.3575
#> 22: 2021-05-18 286.8799
#> 23: 2021-05-19 288.1644
#> 24: 2021-05-20 285.8161
#> 25: 2021-05-21 286.1500
#> 26: 2021-05-22 286.1500
#> 27: 2021-05-23 286.6095
#> 28: 2021-05-24 284.2471
#> 29: 2021-05-25 284.8744
#> 30: 2021-05-26 286.8945
#> 31: 2021-05-27 286.7803
#>           date     rate

get_exchange_rates("USD", "EUR", start_date = "2021-04-18", end_date = "2021-05-18")

#>           date     rate
#>  1: 2021-04-18 0.835317
#>  2: 2021-04-19 0.830670
#>  3: 2021-04-20 0.831007
#>  4: 2021-04-21 0.830900
#>  5: 2021-04-22 0.832272
#>  6: 2021-04-23 0.826549
#>  7: 2021-04-24 0.826641
#>  8: 2021-04-25 0.826986
#>  9: 2021-04-26 0.827641
#> 10: 2021-04-27 0.827118
#> 11: 2021-04-28 0.824243
#> 12: 2021-04-29 0.824961
#> 13: 2021-04-30 0.831981
#> 14: 2021-05-01 0.831950
#> 15: 2021-05-02 0.831463
#> 16: 2021-05-03 0.829221
#> 17: 2021-05-04 0.832241
#> 18: 2021-05-05 0.832902
#> 19: 2021-05-06 0.829121
#> 20: 2021-05-07 0.822149
#> 21: 2021-05-08 0.822149
#> 22: 2021-05-09 0.821853
#> 23: 2021-05-10 0.823757
#> 24: 2021-05-11 0.823247
#> 25: 2021-05-12 0.828181
#> 26: 2021-05-13 0.827931
#> 27: 2021-05-14 0.823423
#> 28: 2021-05-15 0.823323
#> 29: 2021-05-16 0.823314
#> 30: 2021-05-17 0.822800
#> 31: 2021-05-18 0.818073
#>           date     rate